requirements.txt
torch
transformers==4.46.3
yfinance
statsforecast
mlforecast
lightgbm
pandas
numpy
pyyaml
matplotlib
scikit-learn
